requirements.txt
gspread==6.0.1
loguru==0.7.0
oauth2client==4.1.3
python-dotenv==1.0.1
python-telegram-bot==20.8
tweepy==4.14.0
urllib3==1.26.9
